fix(trip-form): move mode settings up one directory
