+++ packages/trip-form/modeSettings.yml
