+++ packages/trip-form/src/index.ts
@@ -11,7 +11,7 @@ import SubmodeSelector from "./SubmodeSelector";
 import MetroModeSelector from "./MetroModeSelector";
 import { useModeState } from "./MetroModeSelector/utils";
 import TripOptions, { Styled as TripOptionsStyled } from "./TripOptions";
-import defaultModeSettings from "./modeSettings.yml";
+import defaultModeSettings from "../modeSettings.yml";
 
 export {
   CheckboxSelector,
